TINY-14067: Use anchor positioning in tooltip component (#11004)
* TINY-14067: Fix local test script to close docker on ctrl+c

* TINY-14067: Deleted custom position logic from the tooltip component, use anchor positioning instead.

* TINY-14067: Move the chat fix for tooltips into the central tooltip anchor class. Fix CSS rule issue with anchor positions.

* TINY-14067: Optimise LESS a bit. Make tooltip story title case like all the rest.

* TINY-14067: Regenerate the screenshot

* Update Tooltip.tsx

* TINY-14067: Simplified null checks

* TINY-14067: Manage docker build return code

* TINY-14067: Hide before adjusting the display property

* TINY-14067: Simplify null checks in effect timeouts. Switch to manual popover to solve weird mouse glitches.

+++ modules/oxide-components/visual-regression-test.mjs
@@ -1,25 +1,49 @@
-import { exec, execSync } from 'child_process';
+import { spawn, spawnSync } from 'child_process';
 
 // This script is used to run visual regression tests in a Docker container.
 
-const devServer = exec('yarn start --ci');
-
-execSync(`docker build -t oxide-components-test-visual . -f Dockerfile.playwright-test`, {
+const devServer = spawn('yarn', ['start', '--ci'], {
   stdio: 'inherit'
 });
+
+const buildResult = spawnSync('docker', ['build', '-t', 'oxide-components-test-visual', '.', '-f', 'Dockerfile.playwright-test'],
+  { stdio: 'inherit' }
+);
+
+if (buildResult.error ?? (buildResult.status ?? 1) !== 0) {
+  devServer.kill();
+  throw buildResult.error ?? new Error(`Failed to build Docker image, return code ${(buildResult.status ?? 1)}`);
+}
 
 const pwd = process.cwd();
 const parentDir = pwd + '/../../';
 
-execSync(`docker run --rm \
- --name=oxide-components-test-visual \
- --userns=host \
- -e TEST_BASE_URL='http://host.docker.internal:6006' \
- -e CI=true \
- -v ${parentDir}:/tinymce \
- -v ${pwd}/src/test/ts/visual.spec.ts-snapshots:/tinymce/modules/oxide-components/src/test/ts/visual.spec.ts-snapshots \
- oxide-components-test-visual ${process.argv[2] === 'update' ? '--update-snapshots' : ''}`, {
+const dockerArgs = [
+  'run',
+  '--rm',
+  '--name=oxide-components-test-visual',
+  '--userns=host',
+  '-e', "TEST_BASE_URL=http://host.docker.internal:6006",
+  '-e', 'CI=true',
+  '-v', `${parentDir}:/tinymce`,
+  '-v', `${pwd}/src/test/ts/visual.spec.ts-snapshots:/tinymce/modules/oxide-components/src/test/ts/visual.spec.ts-snapshots`,
+  'oxide-components-test-visual'
+];
+
+if (process.argv[2] === 'update') {
+  dockerArgs.push('--update-snapshots');
+}
+
+const docker = spawn('docker', dockerArgs, {
   stdio: 'inherit'
 });
 
-devServer.kill();
+docker.on('exit', (code) => {
+  devServer.kill();
+  process.exit(code ?? 0);
+});
+
+process.on('SIGINT', () => {
+  docker.kill();
+  process.exit();
+});
